+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Show "Group Members Only" label for group events [#3912](https://github.com/rokwire/illinois-app/issues/3912).
 ### Changed
 - Updated display strings when linking event to a group, polished GroupsSelectionPopup UI [#3905](https://github.com/rokwire/illinois-app/issues/3905).
+- Make group event to have the same UI as regular event [#3914](https://github.com/rokwire/illinois-app/issues/3914).
 ### Added
 - Show published status in event detail panel for admins only [#3908](https://github.com/rokwire/illinois-app/issues/3908).
 - Added published info link in Events filters panel [#3886](https://github.com/rokwire/illinois-app/issues/3886).